Rework image upload controller for express. Move most logic to service where it belongs. Also adapt e2e tests to express.
