+++ package.json
@@ -69,6 +69,7 @@
     "@nestjs/testing": "^9.1.2",
     "@types/bcrypt": "^5.0.0",
     "@types/jest": "^29.1.1",
+    "@types/multer": "^1.4.7",
     "@types/node": "^18.7.23",
     "@types/passport-jwt": "^3.0.7",
     "@types/supertest": "^2.0.12",
