+++ src/crags/entities/image.entity.ts
@@ -23,6 +23,11 @@ export enum ImageType {
   PROFILE = 'profile',
 }
 
+export enum ImageParentEntityType {
+  CRAG = 'crag',
+  ROUTE = 'route',
+}
+
 export const sizes = [300, 600, 1040]; // Add to this list if other sizes needed by FE (if changed, should reprocess all images)
 
 @Entity()
