+++ src/crags/services/images.service.ts
@@ -1,12 +1,24 @@
 import { InjectRepository } from '@nestjs/typeorm';
 import { Repository } from 'typeorm';
 import { User } from '../../users/entities/user.entity';
-import { CreateImageDto } from '../dtos/create-image.dto';
-import { Image } from '../entities/image.entity';
+import { UploadImageDto } from '../dtos/upload-image.dto';
+import { Crag } from '../entities/crag.entity';
+import { Image, sizes } from '../entities/image.entity';
+import { Route } from '../entities/route.entity';
+import { CragsService } from './crags.service';
+import { RoutesService } from './routes.service';
+import sharp from 'sharp';
+import path from 'path';
+import { env } from 'process';
 
 export class ImagesService {
+  maxSize = { width: 6000, height: 6000 };
+  targetSizes = sizes;
+
   constructor(
     @InjectRepository(Image) private imagesRepository: Repository<Image>,
+    private routesService: RoutesService,
+    private cragsService: CragsService,
   ) {}
 
   async findOneByPath(path: string): Promise<Image> {
@@ -40,36 +52,134 @@ export class ImagesService {
   }
 
   async createImage(
-    createImageDto: CreateImageDto,
+    uploadImageDto: UploadImageDto,
+    imageFile: Express.Multer.File,
     currentUser: User,
-  ): Promise<Image> {
-    const {
-      path,
-      extension,
-      aspectRatio,
-      maxIntrinsicWidth,
-      type,
-      title,
-      description,
-    } = createImageDto;
+  ) {
+    const { entityType, entityId, type, title, description } = uploadImageDto;
+    const inputExtension = path.extname(imageFile.originalname);
+
+    let crag: Crag;
+    let stemBase: string;
+    let parentEntity: Route | Crag;
+    switch (entityType) {
+      case 'route':
+        const route = await this.routesService.findOneById(entityId);
+        crag = await route.crag;
+        stemBase = crag.slug + '-' + route.slug;
+        parentEntity = route;
+        break;
+      case 'crag':
+        crag = await this.cragsService.findOneById(entityId);
+        stemBase = crag.slug;
+        parentEntity = crag;
+        break;
+      // Add other cases when/if adding support for other entity types
+    }
+
+    const stem = await this.generateUniqueStem(entityType, stemBase);
+
+    // Resize image, generate other image sizes, save to disk and retreive some image data needed for FE
+    const { maxIntrinsicWidth, aspectRatio } = await this.processImage(
+      imageFile.buffer,
+      stem,
+      inputExtension,
+      entityType,
+    );
+
+    // Save image info to db
     const image = this.imagesRepository.create({
-      path,
-      extension,
+      path: `${entityType}s/${stem}`,
+      extension: inputExtension,
       aspectRatio,
       maxIntrinsicWidth,
       type,
       title,
       description,
     });
-    if (createImageDto.route) {
-      image.route = Promise.resolve(createImageDto.route);
-    }
-    if (createImageDto.crag) {
-      image.crag = Promise.resolve(createImageDto.crag);
-    }
+
+    // image.crag or image.route ...
+    image[<string>entityType] = Promise.resolve(parentEntity);
+
     image.user = Promise.resolve(currentUser);
 
     await this.imagesRepository.save(image);
-    return image;
+
+    return image.id;
+  }
+
+  private async generateUniqueStem(entity: string, stemBase: string) {
+    let counter = 0;
+    let stem = stemBase;
+
+    while (await this.findOneByPath(`${entity}s/${stem}`)) {
+      counter++;
+      stem = `${stemBase}-${counter}`;
+    }
+    return stem;
+  }
+
+  private async processImage(
+    imageBuffer: Buffer,
+    stem: string,
+    extension: string,
+    entity: string,
+  ) {
+    // Load image from buffer
+    const shImage = await sharp(imageBuffer);
+
+    await shImage
+      .resize({
+        width: this.maxSize.width,
+        height: this.maxSize.height,
+        fit: 'inside',
+        withoutEnlargement: true,
+      })
+      .toFile(`${env.STORAGE_PATH}/images/${entity}s/${stem}${extension}`);
+
+    // Generate all of the predefined sizes and save them as webp, avif and jpeg
+    this.targetSizes.forEach(async size => {
+      // Resize image (only if big enough)
+      const resizedImage = shImage.clone().resize({
+        width: size,
+        height: this.maxSize.height,
+        fit: 'inside',
+        withoutEnlargement: true,
+      });
+
+      // Convert to different formats, suitable for FE
+      const webpImage = resizedImage.clone().toFormat('webp');
+      const avifImage = resizedImage.clone().toFormat('avif');
+      const jpegImage = resizedImage.clone().toFormat('jpeg', {
+        quality: 80,
+        chromaSubsampling: '4:4:4',
+        mozjpeg: true,
+      });
+
+      // Save image of current size in all formats to disk
+      jpegImage.toFile(
+        `${env.STORAGE_PATH}/images/${size}/${entity}s/${stem}.jpg`,
+      );
+      webpImage.toFile(
+        `${env.STORAGE_PATH}/images/${size}/${entity}s/${stem}.webp`,
+      );
+      avifImage.toFile(
+        `${env.STORAGE_PATH}/images/${size}/${entity}s/${stem}.avif`,
+      );
+    });
+
+    // Aspect ratio can be determined from original image metadata, because it will not change. From that also actual max intrinsic width can be determined
+    const { width, height } = await shImage.metadata();
+    const aspectRatio = width / height;
+    const maxIntrinsicWidth = Math.round(
+      Math.min(
+        width,
+        this.maxSize.width,
+        this.maxSize.height * aspectRatio,
+        Math.max(...this.targetSizes),
+      ),
+    );
+
+    return { maxIntrinsicWidth, aspectRatio };
   }
 }
