+++ test/e2e/activity.e2e-spec.ts
@@ -2,10 +2,6 @@ import request from 'supertest';
 import { Test } from '@nestjs/testing';
 import { AppModule } from '../../src/app.module';
 import { QueryRunner } from 'typeorm';
-import {
-  FastifyAdapter,
-  NestFastifyApplication,
-} from '@nestjs/platform-fastify';
 import { MailService } from '../../src/notification/services/mail.service';
 import { initializeDbConn, prepareEnvironment, seedDatabase } from './helpers';
 import { UsersModule } from '../../src/users/users.module';
@@ -15,9 +11,10 @@ import {
   Activity,
   ActivityType,
 } from '../../src/activities/entities/activity.entity';
+import { INestApplication } from '@nestjs/common';
 
 describe('Activity', () => {
-  let app: NestFastifyApplication;
+  let app: INestApplication;
   let queryRunner: QueryRunner;
 
   let mockData: any;
@@ -34,15 +31,9 @@ describe('Activity', () => {
       .useValue(mailService)
       .compile();
 
-    app = moduleRef.createNestApplication<NestFastifyApplication>(
-      new FastifyAdapter(),
-    );
+    app = moduleRef.createNestApplication();
 
     await app.init();
-    await app
-      .getHttpAdapter()
-      .getInstance()
-      .ready();
 
     const conn = await initializeDbConn(app);
     queryRunner = conn.createQueryRunner();
